packages/web-server/requirements.txt
django==3.0.3
psycopg2>=2.7,<3.0
djangorestframework==3.11.0
PyJWT==1.7.1
djangorestframework_simplejwt==4.5.0
django-cors-headers==3.2.1
Pillow==7.0.0
urllib3==1.25.8
sib_api_v3_sdk==6.1.0
gunicorn==19.9.0
elasticsearch==7.8.0
numpy==1.19.1
opencv-python-headless==4.3.0.38
django-paranoid==0.8
nltk==3.5
django-crontab==0.7.1